refactor(at_client_flutter): seperated keychain_manager functoinality into keychain_storage for better testing and cleaner code
