+++ packages/at_client_flutter/lib/src/keychain/keychain_io_impl.dart
@@ -1,53 +1,54 @@
 import 'package:at_auth/at_auth.dart' show KeyIOMixin, WrittenAtKeysIo, AtKeys;
 import 'package:at_client_flutter/at_client_flutter.dart';
-import 'package:at_client_flutter/src/atsign_key.dart';
-import 'package:at_utils/at_logger.dart';
 import 'package:flutter/cupertino.dart';
 
+/// Implementation of WrittenAtKeysIo using Keychain storage
+/// Uses KeyChainManager to perform the CRUD operations related to keychains
+///
+/// This is the main class to interact with keychain for storing and retrieving AtKeys
 class KeychainAtKeysIo extends WrittenAtKeysIo with KeyIOMixin {
   @visibleForTesting
   KeyChainManager keychainManager;
-  final _logger = AtSignLogger('KeychainAtKeysIo');
-  KeychainAtKeysIo(this.keychainManager);
+  KeychainAtKeysIo({KeyChainManager? keychainManager})
+      : keychainManager = keychainManager ?? KeyChainManager();
 
   @override
-  Future<AtsignKey> read(String atSign) async {
-    final atsignKey = await keychainManager.readAtsign(name: atSign);
+  Future<AtKeys> read(String atSign) async {
+    final atsignKey = await keychainManager.getAtSign(name: atSign);
     if (atsignKey == null) {
       throw AtKeyException('AtsignKey not found in keychain for atSign: $atSign');
     }
     return atsignKey;
   }
 
-  Future<List<AtsignKey>> readAll() async {
-    final atsigns = await keychainManager.readAtsigns();
+  Future<List<AtKeys>> readAll() async {
+    final atsigns = await keychainManager.getAllAtSigns();
     return atsigns;
   }
 
   @override
   Future<void> write(String atSign, AtKeys atKeys) async {
-    var atSignItem = await keychainManager.readAtsign(name: atSign) ?? AtsignKey(atSign: atSign);
-    atSignItem = atSignItem.copyWithAtKeys(atSign, atKeys);
-    await keychainManager.storeAtSign(atSign: atSignItem);
+    var atKeys = await keychainManager.getAtSign(name: atSign) ?? AtKeys();
+    atKeys.metadata['atsign'] = atSign;
+    await keychainManager.putAtSign(atKeys: atKeys);
   }
 
-  Future<void> writeToEnrollmentStore(String atSign, String data) async {
-    final store = await keychainManager.getEnrollmentStorage(atSign);
-    await keychainManager.writeDataToStore(store: store, data: data);
+  Future<String?> readEnrollmentFromKeychain(String atsign) async {
+    return await keychainManager.readFromEnrollmentStore(atsign);
   }
 
-  Future<String?> readFromEnrollmentStore(String atSign) async {
-    final store = await keychainManager.getEnrollmentStorage(atSign);
-    return await keychainManager.readDataFromStore(store: store);
+  Future<void> writeEnrollmentToKeychain(String atsign, String enrollmentId) async {
+    await keychainManager.writeToEnrollmentStore(atsign, enrollmentId);
   }
 
   Future<void> deleteEnrollmentStore(String atSign) async {
-    final store = await keychainManager.getEnrollmentStorage(atSign);
-    await store.delete();
+    await keychainManager.deleteEnrollmentStore(atSign);
   }
 
+
+  // Should we support legacy backup format?
   Future<Map<String, String>> getEncryptedKeys(String atsign) async {
-    AtsignKey? atsignKeyData = await keychainManager.readAtsign(name: atsign);
+    AtKeys? atsignKeyData = await keychainManager.getAtSign(name: atsign);
 
     if (atsignKeyData == null) {
       throw AtClientException.message("Failed to fetch the keys for the atsign: $atsign");
@@ -83,40 +84,5 @@ class KeychainAtKeysIo extends WrittenAtKeysIo with KeyIOMixin {
       encryptedAtKeysMap[BackupKeyConstants.APKAM_ENROLLMENT_ID_FROM_FILE] = atsignKeyData.enrollmentId!;
     }
     return encryptedAtKeysMap;
-  }
-
-  /// Function to save keys for the atsign passed to keychain
-  Future<bool> storeAtKeysToKeychain(String atsign, AtKeys atKeys) async {
-    final internalAtClientData = await keychainManager.readAtClientData(useSharedStorage: false);
-    final useSharedStorage = internalAtClientData?.config?.useSharedStorage ?? false;
-    final atClientData = await keychainManager.readAtClientData(useSharedStorage: useSharedStorage);
-    try {
-      final atsigns = atClientData?.keys ?? [];
-      final index = atsigns.indexWhere((element) => element.atSign == atsign);
-      if (index >= 0) {
-        atsigns[index] = atsigns[index].copyWithAtKeys(atsign, atKeys);
-      } else {
-        atsigns.add(AtsignKey(atSign: atsign).copyWithAtKeys(atsign, atKeys));
-      }
-      atClientData?.keys = atsigns;
-      if (atClientData != null) {
-        await keychainManager.saveAtClientData(data: atClientData, useSharedStorage: useSharedStorage);
-        return true;
-      } else {
-        return false;
-      }
-    } catch (e, s) {
-      _logger.severe('exception in storeCredentialToKeychain :${e.toString()}');
-      _logger.severe(s);
-      return false;
-    }
-  }
-
-  /// Function to clear all entries from keychain
-  Future<void> deleteAllAtSignFromKeychain() async {
-    var atsigns = await keychainManager.readAtsigns();
-    for (var element in atsigns.map((e) => e.atSign).toList()) {
-      await keychainManager.deleteAtSignFromKeychain(element);
-    }
   }
 }
